allow updating pages to allow youTube urls

+++ app/Http/Controllers/PageController.php
@@ -71,11 +71,12 @@ class PageController extends Controller
     public function update(UpdatePageRequest $request, Page $page): Redirector|RedirectResponse|Application
     {
         if ($request->hasFile('image')) {
-            if (Storage::exists($page->image_path)) {
+            if ($page->image_path && Storage::exists($page->image_path)) {
                 Storage::delete($page->image_path);
             }
             $image = $request->file('image')->storePublicly('book/'.$page->book->slug);
             $page->image_path = $image;
+            $page->video_link = null;
         }
 
         if ($request->has('content')) {
@@ -86,7 +87,13 @@ class PageController extends Controller
             $page->book_id = $request->book_id;
         }
 
-        if ($request->has('video_link')) {
+        if ($request->has('video_link') && !is_null($request->video_link)) {
+            if ($page->image_path){
+                if (Storage::exists($page->image_path)) {
+                    Storage::delete($page->image_path);
+                }
+                $page->image_path = "";
+            }
             $page->video_link = $request->video_link;
         }
